Add pre and post hooks to CLI Framework
We need a way to run code after the command has run but before we exit.
This requests introduces the concept of 'pre' and 'post' hooks which are
managed by a Runnable instance (program or command).

- Use `{Runnable}.runHooks` and `{Runnable}.hook(type, fn)` instead of
  `{Runnable}.runMiddleware` and `{Runnable}.use`

This change is needed to properly clean up our connection to the Daemon.

Related #98