Introduce Daemon Start Backoff
Instead of waiting for a static second during first run of the daemon we
now utilize an exponential backoff starting at 100ms.

This way, if you have a fast machine you won't have to wait around. We
will need to play with the numbers over time to find the right values.

Closes #112