Introduce Docker for CLI
This pull request introduces a Dockerfile for building and testing the
CLI and Daemon code bases contained in this repository. It leverages two
volumes for bringing code into the container and for outputting build
artifacts (things we can provide to customers).

The build pipeline for local development and running tests is complete
while the ground work for the packaging pipeline has been laid.

*Build Pipeline and Artifact Building*:
Related arigatomachine/cli#131
Related arigatomachine/cli#92

*Running in CI+Testing*:
Closes arigatomachine/cli#142
Closes arigatomachine/cli#120
Closes arigatomachine/cli#95
Closes arigatomachine/cli#148