Implement Login Middleware
There are two new middlewares:

- `session` in `lib/middleware/session` that creates a session object
- `auth` in `lib/middleware/auth` which prints an error and exits if the
  session object is not set on `ctx`.

The session middleware runs on every single command invocation while the
auth middleware only runs on commands that require authentication
information.

This builds on-top of the ability for middleware to early exit and
discontinue the call chain of functions.

All tests have been updated and the underlying functionality has been
completed.

Closes arigatomachine/cli#162