Introduce Project Context to the CLI
This commit introduces the very first version of Context to the CLI. A
link is created using `ag init` to initalize a link between the
directory location and an org, project, service, and environment.

These links are kept inside the ~/.arigato/map.json file. When a
modification is made to the file a lock is acquired. If the lock cannot
be acquired then we error.

All relevant commands have been updated to use the new target middleware
and Target object.

Big changes have been made to `ag status` to now print the complete
operating context.

Closes arigatomachine/cli#328
Closes arigatomachine/cli#327
Closes arigatomachine/cli#326
Closes arigatomachine/cli#325
Closes arigatomachine/cli#324