Propagate exit code from 'ag run'
Commands can now pass a boolean (which gets mutated into a 0 or 1), or
an integer between 0 and 127 as the exit code. This exit code is then
propagated back up the chain to `bin/arigato` which calls
`process.exit`.

If a post-hook middleware does not succeed then it's exitCode takes
precedence.

Closes arigatomachine/cli#296