Introduce Prefs and Arigato Public Key
As a part of arigatomachine/cli#451 we introduced a preference to
override the default arigato root public key to use for verifying the
integrity of the key claim tree.

The daemon now reads in this root public key from the distributed
bundle. If the override is defined in the `.arigatorc` file it uses that
file instead.

If the file does not exist OR the file is invalid the daemon errors and
will not start.

Related arigatomachine/cli#451