Introduce 'ag invites:send' for emailing an invite
A user can now send an invite to their organization to an email address.
They can also specify which teams to join using the `--team` flag which
can be provided many teams for multiple teams.

The members team is implicitly added to the pending_teams list such that
the user does not have to specify it manually.

Related arigatomachine/cli#569