Introduce ag invites:accept command
The invite acceptance command allows a user to accept an invite from a
logged in state. If they are not logged in, then they are able to create
an account, or login to their pre-existing accounts.

If they are creating an account then they don't need to verify their
email address.

Related arigatomachine/cli#569