Move remaining secret set/unset logic into daemon
Determining if there is a previous credential is now the responsibility
of the Daemon instead of the CLI. This lays the final groundwork for
integrating the `CredentialTree` with the Daemon and introducing Keyring
and Keyring Members.

Opened arigatomachine/cli#620 to re-add logic to prevent unsetting a
secret twice.

Related arigatomachine/cli#529