Introduce ag keypairs and keypairs:generate
`ag keypairs` allows the user to view all of their keypairs, or just
keypairs for a specific org. If they are in a linked directory then they
can use `-a, --all` to list *all* organizations.

`ag keypairs:generate` allows the user to generate a keypair for a
specific organization. They can also use `--all` to generate keypairs
for any organizations where they are missing valid keypairs!

Closes arigatomachine/cli#524