Fix login and run bug using ag run
If you were to login at the same time as you `ag run` (e.g. `AG_EMAIL=x
AG_PASSWORD=y ag run bash test.sh`) the cli would login but crash
attempting to spawn.

The reason is that we were setting the `ctx.params` to `[]` when you
tried to login! Silly side-effects!

Add a nice error to prevent this from happening in the future *and*
remove the offending side-effect.

Closes arigatomachine/cli#779