Ignore stale pid files
Treat stale pid files the same as if the daemon had happily exited.
This will allow commands that start the daemon to not exit out when a
stale pid exists, and instead start the daemon.

Also, fix a potential race condition in the daemon itself, If one daemon
is stopping while the other is starting.

closes #842