Support number types for credential values
It would be nice to expose the values as ints or floats, but the string
representation is enough for now.

unset, string, and number should cover all existing values. The minimist
module would convert cli options to numbers if they looked like numbers,
but did not change any other types.

closes #899