Introduce credentialGraphSet for answering CredentialGraph questions
credentialGraphSet holds credentialGraph versions, and provides methods
for ansering questions about the credentialGraphs related to credential
or version reachability/liveliness.

For the first method, implement HeadCredential, which finds the most
recent version of a Credential with the given PathExp and Name.

Use HeadCredential when adding credentials to a keyring to find their
previous.