Filter Credentials by matching pathexp and name
As a result of arigatomachine/cli#1107 and arigatomachine/registry#323
the daemon must perform name and pathexpression fitlering within a
keyring on its own.

This responsibility use to live in the server but has been moved to the
daemon to fix a bug. A secondary benefit is that it further aligns the
daemon with it's long term responsibility (acting as a distributed
database).

Closes arigatomachine/cli#1107