Introduce the worklog command
`worklog` discovers and tracks important tasks to do within `torus`. The
first type of item it tracks are credentials that should be rotated due
to users being removed from an organization.

`worklog resolve` is included too, but is hidden. I started writing this
including a worklog item for keyring version bumping, which the worklog
could resolve for you, but you can't actually bump a keyring version
without appending a credential (and there would't be any point in that
regardless).

closes #480