Support Machine and User sessions
In order to support the concept of users, we have to separate out
`identity` and `authentication`. Who someone is, versus, what they use
to prove they have access to assume that identity (and the associated
roles).

Therefore, I've reworked the Session object to have type (not logged in,
user, and machine) which have an `identity` and `auth` field.

A user's identity and auth field are the exact same; a user object. A
machine's is not, it's the Machine and MachineToken object.