Primitives and PDPKA
Primitives for MachineToken and Machine!

The LoginKeypair object is used for performing passphrase-dervied public
key authentication which creates a signing keypair derived from a
section of the scrypt output based on the given passphrase.

We then sign a logn_token with our public key instead of deriving an
hmac (which is currently done for users).

This introduces logic for generating the PublicKey component of the
MachineToken (and other changes related to the introduction of
MachineTokens).