Introduce 'machines create' command
The brand spanking new `machines create` command which allows you to
create a machine, specify it's team (role), generates its tokens (and
underlying keypairs), uploads these to the server, and then displays the
token values to be copied.

This commit only contains changes directly related to introducing the
create command.