Add machines into newly created keyrings
Just like for users, when we need to create a new keyring, ensure that
all machines that should have access to the keyring do in fact have
access.

As we do for users, for now, add *all* machines in the org, and rely on
server-side ACL blocking for access restriction.