Include a systemd service with our rpm
The service can be configured with machine token credentials in
/etc/torus/token.environment, for use in production environments.

Other processes running on the system can access secrets from torus by:
- running as a user in the `torus` group
- setting the TORUS_ROOT environment variable to /var/run/torus

While there is a torus user, this user should *not* be used for running
processes that need access to secrets. It is an isolated sandbox that
has access to the machine's token credentials.