Introduce torus template command
The `torus template` command allows a user to inject secrets stored in
torus as they derive a configuration file from a template.

The current templating engine is the [raymond
handlebars](https://github.com/aymerick/raymond) engine for Go
which supports v3 of the handlebars spec.

Also includes bumps to our dependencies, as I had to include the raymond
library. The interface for `readline.NewCancelableStdin` has changed in
newer versions which I've upgraded us to.

Closes manifoldco/torus-cli#51