Introduce daemon side credential pruning
credentialGraphSet.Prune returns all CredentialGraphs within the
credentialGraphSet that are still active (they have reachable
credentials). Further, it only returns the credentials in each graph
that are still active.

Use Prune for view operations, so we can be sure that even before we
decrypt credentials, we are only returning the active version of a
credential.

Closes #98