Observer: put a Read/Write mutex in place.
`*notifier.Notify()` can be used concurrently. Because of this behaviour, it can
trigger a race condition when two events happen at the same time where
one would incremment the current counter and one would read from it.

This implements a Read/Write lock which allows this behaviour by locking
the instance on updating the counter and only adding a Read lock when
reading the value. This allows us to only write from a single point at a
single moment in time, but read from multiple points at a single moment
in time. It will also prevent users from reading the counter when
someone else is writing to it.