ObserverTest: prevent a deadlock situation.
The tests described in the test file didn't really describe a real world
situation. In the real world, an observer is either always present and
started well before we send notifications or no observer will be ever
present at all.

Running the tests, a deadlock occured sometimes, locking up the tests
and eventually failing them. This was because the Observer in our tests
didn't register itself in time before we created a new notification. The
notification then got sent through the system and dropped because no
observer was present. Once the observer then started, it would keep listening
until the notification came through, resulting in a deadlock since the
event was already dropped beforehand.

This patch does actually not fix the deadlock situation itself, it is
still possible to happen - there is however logging present now to
document this. We do however set up the tests more reliable in a way
that we expect the real world to behave. We ensure that the observer is
registered before we send the notification. This way once we send a
notification, the observer sends it pack to the ResponseRecorder.