Observer: handle locks accordingly.
The panic is caught higher up the chain and logged out that way. To not
have the lock persist, we'll start a reading lock, read out the data we
need to read and the unlock the reading lock. This way, in case a panic
happens the lock is actually properly unlocked.