Client: send data to the channel instead of close.
Calling close is none blocking, it closes the channel immediately and
reads will still happen from the `stream.Events` channel instead of the
`<- done` channel.

We want to block here until the message is actually read from the done
channel so we can continue processing the end result and not show other
event messages.