Put worklog item type logic into their own types
Extract out the `List` and `Resolve` logic for each worklog item type
into new structs that implement a common `handler` interface.

This keeps the logic for each type distinct, which should (hopefully)
make for easier reading, and easier addition of new item types.

With the logic now distinct and behind a map lookup, take advantage of
this during `Get` and `Resolve` to extract the item type from the
requested ID, and only lookup that type of item, bypassing all the
others.