APIClient: listen on done channel on error.
When an error happens in unmarshaling the JSON, we'll end up in a
deadlock. This is because we do a return, ending the for and select
loop, which listens on the done channel. This means that nothing will
ever get read from the done channel. On line 166, we try to put
something on the done channel and block until it's read.

This goroutine acts as a reader for that channel and will unblock our
message, allowing us to drop the incoming events in case of an
unmarshal error.