Skip keyless users when building keyrings
If a user does not have an unrevoked encryption key when we're building
a keyring, skip over them instead of blowing up. This way, other users
can use the keyring values, and the worklog can warn let people know to
encode this user in the keyring once they have generated keypairs.