Stop attempting to convert credentials to numbers
We support types for credentials, which is nice. However, we can't
accurately know if a string given to us on the command line that *looks*
like a number really should be a number.

If we assume it should be a number, we end up losing data, like leading
zeros, or converting long decimals to scientific notation.

For values given on the command line, stop attempting to convert the
value, and instead store it as a string.

Closes #163