Use a common interface for registry API calls
Introduce a RoundTripper (in both `daemon/registry`, and `api` pkgs) that
exposes NewRequest and Do methods. Use it instead of Client.
This interface is identical to what `daemon/registry` was already using,
but is different in `api`, so include a shim. Convert all types in `api`
that make purely proxied calls to this new interface.

This step will simplify unification of the two groups of apis.