Disable log output for normal command line use
We only use logging in the daemon side of things. For the cli frontend,
we centralize output in the cmd package.

So that we can reuse the registry api endpoints, and still log errors on
the daemon side, keep the log statements in, but shut down logging for
command line usage.