Populate rich worklog details
For each worklog item type, attach a Details struct containing type
specifid information (such as teams pending for an invite, etc).

Use some of this data for generating the item subject and summary string
in the cli, rather than the daemon.

This data will be used for presenting extra data in the worklog, and
formatting it in a most joyful manner.