Makefile: set up to build binaries for Windows.
Because we want to use `torus` instead of `torus-cli`, we need to
specify our own output (`-o`) binary name. This leads to the Windows
version not automatically being appended with `.exe`.

This replaces the `OUT` variable used in the makefile to check if we've
provided an `OUT` option ourselves - if we want to crosscompile for
windows for example. If so, we'll use this variable. If not, we'll check
if we're building for Windows and in the case we are, we'll use
`torus.exe`. If not, we'll use `torus`.

This also sets up the release targets to build a Windows version.