cmd: profiles: test: Fetch all teams to which the specified user belongs
The first step in testing a user's access to a specific path is to get
all the groups the user belongs to. This requires three backend API
calls:
- Profiles.ListByName on the user
- Teams.GetByOrg on the org
- Membership.List on each team returned by the above call

I'm assuming each of those API calls can be expensive so they are best
executed in parallel as much as possible. The first two steps are
independent and can indeed be executed in parallel. The last step
depends on the first two so must happen after, but, Membership.List must
be called on each team, and this can be done in parallel.

For the most part I followed the existing approach of using a WaitGroup
and closures. I deviated from this a little when filtering teams by the
user, where I used a channel to send the final list of teams back to the
caller. This has one disadvantage in that errors encountered in the
goroutine writing to the channel can't easily be propagated back to the
caller (see TODO).