cmd: allow/profiles: Reuse parseRawPath in allow command
I've moved parseRawPath into allow.go alongside parseAction. I also had
to modify the command to also return the secret since the latter is used
in the body of doCrudl.

One additional difference is that "Invalid path expression" is now
returned as a NewExitError instead of a NewUsageExitError. This is
primarily driven by the use of parseRawPath in policyTouchesPath
predicate; it doesn't make sense to create a UsageExitError there since
any parse errors encountered therein would not be a result of bad user
input. I think this is not so bad since other errors returned by
parseRawPath (and the code it replaced) are not UsageExitErrors but
arguably could be since they are a result of bad user input.

I am amenable to changing parseRawPath to accept a cli.Context so it can
generate a UsageExitError.