Use gometalinter --install to acquire linters
gometalinter recently switched from using go's built in `go vet` tool to
a fork to fix several bugs that their users have experienced. However as
a result this broke the Torus build as we were not using `gometalinter
--install` to bring in the various linters.

To fix this in the short term, I've switched our bootstrap setup to use
`gometalinter --install` and opened an issue for properly vendoring our
bootstrap dependencies (#286).