Vendor build dependencies
Often our build was breaking due to changes in one of our build
dependencies such as `gometalinter`.

To work around this, I've adopted a vendoring strategy for these
binaries which are built into `vendor/bin/$NAME`. Then everything that
requires this dependency in the Makefile runs the binary from that
location.

Closes manifoldco/torus-cli#286