Conversion to manifoldco/promptui
Instead of continuing to use and maintain our promptui library, we've
decided to switch to `manifoldco/promptui` which was pulled out from
torus-cli some time ago.

At the same time, I've added better support for "detached terminal"
states where stdout is not a terminal. In these situations, the user is
not prompted, instead, the provided value (if any) is validated and
checked against known options.

On top of these changes, I've added a wrapper around every command to
ensure that errors are displayed properly.

Several other cosmetic changes have been bundled in to make it easier to
display warnings and other types of information to the user as a part of
the `ui` package.

Closes #303