Add dep prune options to dep manifest
With 0.4.0, dep now runs `prune` automatically when when using `ensure`.
Dep will check in the manifest for what options are needed for pruning.

This commit sets the prune settings to remove non-go files (exlcuding
LICENSE and AUTHORS), go test files. Unused (non-imported) packages need
to be included because of the tools/ package.