No longer warn if not in .git folder
A few components of the Makefile were using `git describe` to figure out
the current tag. However, when the container is built it's done outside
of a git repository leading to a ton of non-important error messages
popping up.

Instead, we only call git describe if we're actually in a `.git` folder.