Fix bug preventing an error from being mutated
In certain circumstances a StatusCode would not be set on an
apitypes.Error. Instead of having to set two properties, I've added a
lookup map for each error type, expanded the error types and added
functionality for converting an error type to a status code nad back.

This should prevent this specific bug from occurring in the future.