Spinners obey core.progress pref and unblock view
We had expoed spinners through a few different apis in the ui layer, as
a result some were obeying the `core.progress` preference while others
were not. I've fixed this by pushing the responsibility for checking
whether a spinner should/is displayed into the actual spinner itself.

At the sametime, I've fixed a bug where it was possible for a spinner to
never be removed due to a panic.

Closes #376 #377