Fix prefs set when not on darwin
If a user attempted to set a preference on a non-darwin operating system
it would fail silently. This is because we were ignoring errors from
`user.Current()` which requires c-go to get the current user.

Since we're not building Torus with c-go (on purpose), we'd get an error
back when we called that method on linux or windows. I've worked around
this by writing a shim and no longer using `user.Current()`.

Closes #375